requirements.txt
pyspark>=3.5
pandas>=2.1
yfinance>=0.2.28
kafka-python>=2.0.2
mysql-connector-python>=8.0
fastapi>=0.104
pytest>=7.4